feat: 팀 삭제/탈퇴 모달 통합
#380
