--- src/containers/team/DeleteTeamModal/type.ts
@@ -1,6 +1,0 @@
-import { Team } from '@/types';
-
-export interface DeleteTeamModalProps extends Pick<Team, 'id' | 'name'> {
-  isOpen: boolean;
-  onClose: () => void;
-}
