+++ src/containers/team/TeamActionModal/type.ts
@@ -1,0 +1,7 @@
+import { Team } from '@/types';
+
+export interface TeamActionModalProps extends Pick<Team, 'id' | 'name'> {
+  isOpen: boolean;
+  onClose: () => void;
+  actionType: 'delete' | 'leave';
+}
